requirements.txt
fastapi==0.120.2
fpdf2==2.8.5
Jinja2==3.1.6
langchain==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.0.10
pydantic==2.12.3
pydantic-settings==2.11.0
python-docx==1.2.0
python-dotenv==1.2.1
uvicorn==0.38.0